Move the logging utility to the logging folder
